google-search-service-v2/requirements.txt
google-adk
google-generativeai>=0.5.0
fastapi
requests
uvicorn
requests
google-cloud-aiplatform
google-api-python-client
deprecated
python-multipart
google-cloud-pubsub
